refactor: DayBar.utils.ts의 스타일 함수를 DayBar.styles.ts로 통합 후 삭제
